refact(SyncWorker): split initial process into a specific file
